Fixed TeleOp, added Autonomous that Goes 1 Rotation
I have fixed an issue concerning the back left motor having it's value
assigned to the value of the back right motor, and divided by 3. This
only happened when the bumpers were not pressed down. There is also an
Autonomous program that goes exactly 1 wheel rotation and will be used
to meausure how far the robot goes for Autonomous purposes.