Added the West Sound STEM Showcase Program
Added the West Sound STEM Showcase Program, a simple program to control
a simple robot with two motors and one servo. You can control the motors
with the right and left joysticks and you can control the servo with the
left and right bumpers. This program only uses one controller.